Generalizations of the cryptography primitives to be used in more than one scheme
