+++ src/crypto/ciphersuite.rs
@@ -1,0 +1,10 @@
+// +++++++++ Generic Ciphersuite +++++++++
+pub enum BytesOrder {
+    BigEndian,
+    LittleEndian,
+}
+
+pub trait ScalarSerializationFormat {
+    fn bytes_order() -> BytesOrder;
+}
+pub trait Ciphersuite: frost_core::Ciphersuite + ScalarSerializationFormat {}
