+++ src/eddsa/mod.rs
@@ -1,5 +1,5 @@
 //! This module serves as a wrapper for Ed25519 scheme.
-use crate::generic_dkg::{BytesOrder, Ciphersuite, ScalarSerializationFormat};
+use crate::crypto::ciphersuite::{BytesOrder, Ciphersuite, ScalarSerializationFormat};
 use frost_ed25519::keys::SigningShare;
 use frost_ed25519::{Ed25519Sha512, VerifyingKey};
